+++ packages/@leight/i18n/src/hook/useRedirect.ts
@@ -1,7 +1,8 @@
-import {useEffect} from "react";
-import {useRouter} from "./useRouter";
+import {type IHrefProps} from "@leight/utils";
+import {useEffect}       from "react";
+import {useRouter}       from "./useRouter";
 
-export const useRedirect = (href: string) => {
+export const useRedirect = (href: IHrefProps) => {
     const {push} = useRouter();
     useEffect(() => {
         push(href);