+++ packages/@leight/i18n/src/ui/Link.tsx
@@ -7,7 +7,7 @@ import {
 import {useRouter} from "../hook/useRouter";
 
 export interface ILinkProps extends ComponentProps<typeof CoolLink> {
-    query?: Record<string, any>;
+    query?: Record<string, any> | undefined | null;
 }
 
 export const Link: FC<ILinkProps> = (
@@ -20,7 +20,7 @@ export const Link: FC<ILinkProps> = (
     return <CoolLink
         href={buildUrl({
             path:        `/${locale}${href}`,
-            queryParams: query
+            queryParams: query || undefined,
         })}
         {...props}
     />;