+++ packages/@leight/utils/src/utils/linkTo.ts
@@ -1,0 +1,8 @@
+import {compile}         from "path-to-regexp";
+import {type IHrefProps} from "../$export/IHrefProps";
+
+export const linkTo = (
+    {
+        href,
+        query
+    }: IHrefProps) => compile(href.replaceAll(/{(.*?)}/g, ":$1"))(query);